pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "bibtex-updater"
dynamic = ["version"]
description = "Replace preprint BibTeX entries with published versions and validate bibliography references"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    { name = "Patrik Reizinger" }
]
keywords = [
    "bibtex",
    "bibliography",
    "arxiv",
    "preprint",
    "crossref",
    "citation",
    "latex",
    "academic",
    "research",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering",
    "Topic :: Text Processing :: Markup :: LaTeX",
    "Typing :: Typed",
]
dependencies = [
    "bibtexparser>=1.4.0",
    "requests>=2.28.0",
    "crossref-commons>=0.0.7",
    "httpx>=0.24.0",
    "rapidfuzz>=3.0.0",
]

[project.optional-dependencies]
scholarly = [
    "scholarly>=1.7.0",
]
zotero = [
    "pyzotero>=1.5.0",
]
organizer = [
    "pyzotero>=1.5.0",
    "pyyaml>=6.0",
]
organizer-claude = [
    "pyzotero>=1.5.0",
    "pyyaml>=6.0",
]
organizer-openai = [
    "pyzotero>=1.5.0",
    "pyyaml>=6.0",
]
organizer-embedding = [
    "pyzotero>=1.5.0",
    "pyyaml>=6.0",
    "sentence-transformers>=2.2.0",
]
all = [
    "scholarly>=1.7.0",
    "pyzotero>=1.5.0",
    "pyyaml>=6.0",
    "sentence-transformers>=2.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pre-commit>=3.6.0",
    "black>=24.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "types-requests>=2.31.0",
    "build>=1.0.0",
    "twine>=5.0.0",
]

[project.scripts]
bibtex-update = "bibtex_updater.cli.updater_cli:main"
bibtex-check = "bibtex_updater.cli.fact_checker_cli:main"
bibtex-filter = "bibtex_updater.cli.filter_cli:main"
bibtex-zotero = "bibtex_updater.cli.zotero_cli:main"
bibtex-zotero-organize = "bibtex_updater.cli.zotero_organizer_cli:main"

[project.urls]
Homepage = "https://github.com/rpatrik96/bibtexupdater"
Documentation = "https://github.com/rpatrik96/bibtexupdater#readme"
Repository = "https://github.com/rpatrik96/bibtexupdater.git"
Issues = "https://github.com/rpatrik96/bibtexupdater/issues"
Changelog = "https://github.com/rpatrik96/bibtexupdater/blob/main/CHANGELOG.md"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
version_scheme = "guess-next-dev"
local_scheme = "no-local-version"

[tool.hatch.build.hooks.vcs]
version-file = "src/bibtex_updater/_version.py"

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
    "/docs",
    "/examples",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bibtex_updater"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312", "py313"]

[tool.ruff]
line-length = 120
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without explicit strict parameter
]

[tool.ruff.lint.isort]
known-first-party = ["bibtex_updater"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
packages = ["bibtex_updater"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/bibtex_updater"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
